Move searchbyisbn.exception to ervinszabolibrary
