+++ src/main/java/com/reka/lakatos/searchbyisbn/crawler/ervinszabolibrary/exception/BookDownloadException.java
@@ -1,4 +1,4 @@
-package com.reka.lakatos.searchbyisbn.exception;
+package com.reka.lakatos.searchbyisbn.crawler.ervinszabolibrary.exception;
 
 public class BookDownloadException extends RuntimeException {
     public BookDownloadException(String message, Throwable cause) {
