+++ src/main/java/com/reka/lakatos/searchbyisbn/crawler/ervinszabolibrary/exception/BookListDownloadException.java
@@ -1,4 +1,4 @@
-package com.reka.lakatos.searchbyisbn.exception;
+package com.reka.lakatos.searchbyisbn.crawler.ervinszabolibrary.exception;
 
 public class BookListDownloadException extends RuntimeException {
     public BookListDownloadException(String message, Throwable cause) {
